Make OpenGL-ES 2.0 front/back-end buildable, and update README
The OpenGL-ES 2.0 front-end and related EGL (X11) back-end now
build without errors when tested on an ARM platform with Mali-400
but a few bugs still need to be eliminated (such as invisible
objects, incorrect text shader, and buggy mouse control). These
bugs are likely to be relatively easy to fix (being introduced
after changes to an older version that was working correctly on
the platform).

The optimized text shader presents some challenges on the OpenGL-ES
2.0 platform because the text string is packed into a uniform byte
array (native representation), which poses problems because the
GLSL variant used in OpenGL-ES 2.0 only supports signed 32-bit
integers and does not support bitwise logical operators nor shift
operators.

Update the README.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>