Improve the light prominence comparison algorithm
Improve the light comparison algorithm (used when the number of
maximum number visible active lights is smaller than the actual
number of visible lights) with calculation of the intensity (based
on light color) and in the case of point source light of the
actual light color intensity at the point of interest.

This may still be suboptimal, for example in case of bright/far-
ranging spot or beam lights (which have low priority), or a small
light that is always very close to the point/object of interest so
may be selected as the most prominent light during single pass
rendering (so that more significant lights are skipped).

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>