Introduce priority-based error and logging message functions
Information and error messages are now printed based on the
priority of the message (e.g warning, information, logging).
The degree of verboseness can be selected by the application
(from quiet to full logging, the latter only if the
'DEBUG_RENDER_LOG flag is set). Fatal errors are also handled
more efficiently by a non-returning function that also raises
an exception that acts as a breakpoint in a debugger.

Only part of the previously existing messages (using printf,
exit etc) has been converted to the new system. The rest
should follow later.

The back-end and the application itself can also make use
these functions.

An option could be added to route messages to the framebuffer
instead of the standard output.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>