Update demo8 and document shadow volume visibility test bugs
Update demo8. Add lights casting shadows through the gratings and remove the
light inside the user-controller object.

Update the README to reflect two shadow volume bugs visible in demo8:
A large grating shadow overlaying the window probably as a result of a
missing darkcap from the shadow volume cast by a corner light. This
is a problem with the shadow volume visibility test, and disabling
the the test fixes it. Additionally, the grating shadow from the
nearby light can disappear; this is also fixed by disable the test.

Another occasional problem is an error message relating to unexpected
number of shadow pyramid sides (shadow volume pyramid cones are
generated for point source lights). This occurs in demo8, and can
ocassionally be seen in demo8.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>