Split and clarify shadow volume visibility tests
The two shadow volume visibility tests (for a shadow volume
cast by object delimited by the light volume, and for the darkcap,
the infinite projection of a pyramid or similarly shaped shadow
volume of a point light source), can now be seperately enabled
and disabled.

The darkcap test is disabled by default, due to bugs. It can be
disabled/enabled using the keyboard interface in the back-end
(keys X and C).

Additionally, this commit also makes working the GUI keyboard
option to force depth-fail stencil shadow volume rendering.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>