Rework and improve geometry scissors; fix spotlight shadow volume
Overhaul the geometry scissors implementation, especially the
geometry scissors cache, adding code to maintain consistency when
some lights are skipped early. This addresses potential errors
when the frustum hasn't changed to existing visibility data is to
be used. When a light has a variable light volume, but has defined
worst case bounds, it is now included in the visibility lists based
on the worst case bounds, and can be skipped when the actual volume
is not visible, while entries in the object geometry scissors
caches for the light will have been reserved when the frustum
changed so that they can be filled in when the light moves into the
frustum.

There is room for optimization, since the solution above may
increase CPU utilization. One improvement would be to detect when
the frustum is changing every frame (common occurence when the
viewpoint is moving), and in that case omit any geometry scissors
cache preparation and storage; only when the frustum becomes
stationary for two or more frames would the scissors caches be
initialized for use in subsequent frames with unchanging frustum.

Some variable name have been changed (e.g. the aray of static
objecst within a static light volume has been renamed from
visible_object to light_volume_object) and some comments edited.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>