Eliminate unnecessary shaders when shadow mapping or HDR is disabled
When certain rendering options, such as shadow mapping and HDR
rendering, are disabled (as is the case for OPENGL_ES2), eliminate
unused code and do not load shaders that will not be used. The
shader lists used when initializing shaders before a lighting pass
during multi-pass rendering is also optimized.

Overall, this optimization should improve performance a little
because of slightly faster code and reduced GPU memory usage
(unused shaders), apart from a reduction in binary code size.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>