Improve handling of NPOT textures, and update textures used
Detect non-power-of-2 texture capabilities of the hardware,
and limit the number of mipmaps to one when necessary.

Streamline the texture loading functions somewhat.

A previously missing small transparent texture in demo7
(when run with OpenGL) has been added.

The non-power-of-two beachball texture has been renamed and
multiple mipmaps provided for platforms that support it.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>