Fix compilation on OpenGL ES-2.0, and tweak NPOT support
Support for mipmapped compressed NPOT textures is enabled for
OpenGL, even though it might not be a good idea generally since
the compressed texture mipmaps will have non-multiple-of-four
dimensions.

Unsurpisingly, this combination doesn't work on the OpenGL ES 2.0
Mali-400 platform, so it is disabled. But a non-mipmapped
compressed texture with a height that is not a multiple of four
seems to work.