Provide a testing and validating program for the RNG
A testing and validation program for the random number generator
module is provided. It does performance tests and validates
the random number functions that have an integer range by
comparing the standard deviation of generated numbers to the
one expected statistically. The CMWC RNG, although fast and
simple, seems to be robust with regard to the statistical
distribution of the random numbers, even with a small state size.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>