Update demo4 and add demo4b
Update demo4 (a navigable highly detailed Earth model) to work better with
the current version of the library. Add demo4b which provides a
choreographed view of a spinning Earth as well as close-up terrain
flyovers, without user control.

The required detailed earth surface color, specularity, night color, and
elevation map textures, which have very large file size, are not provided
in this repository, so demo4/4b are not functional as provided. However,
it is possible to change the code (demo4.cpp) to use different textures,
and texture size and other parameters are configurable, which means demo4
should be be adaptable to other combinations of color and elevation map
textures (with optional specularity and dark side textures) that are laid
out in a longitude/latitude configuration, such as different satellite-
based data sets of the Earth or other planets or moons, or artificially
generated ones.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>