Add support for single-component textures (uncompressed)
Loading of uncompressed single-component (which can be described
as grayscale, red-only, or luminance) textures from a PNG file is
now supported, including uploading to video memory. This should bring
performance improvement when single component textures are used.
Support for single-component texture compression formats still has
to be completed.

Also add support for the loading into system memory, but not
uploading to video memory, of 16-bit color component textures. This
is utilized in demo4(b) to allow for an elevation map with higher
precision (which is used to generate the terrain mesh, and not
directly used on the GPU).

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>