Improve texture handling and, add standard rectangle objects
Improve text handling by fixing single-component transparent
textures, and introducing options to keep uncompressed texture
data in system memory or not to upload the texture to the GPU.

Provide a function that generates a texture from a text string
using a given font stored in system memory. A new function to
create a system memory font has also been added.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>