Major update of back-end interface, increasing flexibility
The back-ends are now integrated into a seperate library that
can contain one or more back-ends (either OpenGL 3.0+ or OpenGL-ES
2.0 back-ends, not mixed). The back-end to be used can be selected
at run-time.

sreBackend and sreApplication classes have been introduced, with
sreBackend defining the low-level back-end-specific functions and
sreApplication covering the customizing of a demo or other program,
with configurable Step() functions, and different modes of user
control. sreApplication and functions to set-up/initialize the
application and run the main loop are included in the back-end
library, making a stand-alone application possible without having
to deal with GUI and other details details.

The optional Bullet physics interface is also integrated into the
back-end.

Some bugs have been fixed, including navigation in demo4 (Earth
terrain demo).

This version is an alpha version of v0.2.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>