Add function to load mipmapped texture while skipping large levels
Add the sreCreateTextureLimitLevel() function to allow the mipmap
level that is actually used to be limited to a given size at the
texture level. Only the mipmaps levels in the allowable size range
are uploaded to the GPU. Previously, detail reduction only happened
in conjunction with global texture detail flags.

To facilitate this, sreTexture::Load() has been added. The
sreCreateTexture() functions are the preferred way to create
textures, while directly invoking the constructor is deprecated
for use applications.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>