Optimize cube map shadow maps
Use a new method to implement point light cube shadow maps using
built-in support for cube maps in the shader language, instead of
calculating everything explicitly. Work in progress, but work with
OpenGL 3+.

Code is in place to support OpenGL-ES 2.0 cube shadow map as well
(when the GL_OES_depth_texture_cube_map extension is present), but
this is untested.

Update bugs/potential improvements in README.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>