Fix specular light calculation, normalize directional lights
In the lighting shader, only take into account intensity of the
light and the object's specular reflection color, and not the
color of the light, when adding the specular contribution.

Make sure directional light directions are normalized in demo
programs. This was causing erratic variation in the intensity of
specular highlights.

Set default for USE_REFLECTION_VECTOR_GLES2 back to NO.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>