Update reflecting changes to the vector classes from DataSetTurbo
Update for changes in operators and syntax for the vector classes
defined in DataSetTurbo (e.g. * operator for Vector3D become
pair-wise multiplication, not dot products, which must be calculated
using Dot(x, y)). This makes the operator syntax consistent with
the shader language.

Some just-in-case cautionary changes to bounds.cpp reflecting concerns
about whether sreBaseModel member functions are really const (the changes
are probably unnecessary, because sreBaseModel does not contain the
affected fields, which are only in sreLODModel).

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>