Add ETC1 and RGTC2 versions of normal map texture used in demo1
Add compressed texture versions of the normal mam texture used in
in demo1. For OpenGL, the texture format is signed RGTC2
(BC5_SNORM), while for OpenGL ES 2.0 (KTX file) the format is
ETC1. The KTX file is currently unlikely to be used because
demo1 is compiled with normal mapping disabled for OpenGL ES 2.0.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>